+++ awscli/customizations/codedeploy/deregister.py
@@ -13,10 +13,11 @@
 
 import sys
 
+from botocore.exceptions import ClientError
+
 from awscli.customizations.commands import BasicCommand
 from awscli.customizations.codedeploy.utils import \
     validate_region, validate_instance_name, INSTANCE_NAME_ARG
-from awscli.errorhandler import ClientError, ServerError
 
 
 class Deregister(BasicCommand):
@@ -132,8 +133,8 @@ class Deregister(BasicCommand):
                         UserName=params.user_name,
                         PolicyName=policy_name
                     )
-        except (ServerError, ClientError) as e:
-            if e.error_code != 'NoSuchEntity':
+        except ClientError as e:
+            if e.response.get('Error', {}).get('Code') != 'NoSuchEntity':
                 raise e
         sys.stdout.write('DONE\n')
 
@@ -148,8 +149,8 @@ class Deregister(BasicCommand):
                         UserName=params.user_name,
                         AccessKeyId=access_key['AccessKeyId']
                     )
-        except (ServerError, ClientError) as e:
-            if e.error_code != 'NoSuchEntity':
+        except ClientError as e:
+            if e.response.get('Error', {}).get('Code') != 'NoSuchEntity':
                 raise e
         sys.stdout.write('DONE\n')
 
@@ -159,7 +160,7 @@ class Deregister(BasicCommand):
         ))
         try:
             self.iam.delete_user(UserName=params.user_name)
-        except (ServerError, ClientError) as e:
-            if e.error_code != 'NoSuchEntity':
+        except ClientError as e:
+            if e.response.get('Error', {}).get('Code') != 'NoSuchEntity':
                 raise e
         sys.stdout.write('DONE\n')